+++ plugin/src/net/groboclown/idea/p4ic/v2/server/connection/ServerConnectedController.java
@@ -24,6 +24,7 @@ import org.jetbrains.annotations.NotNull;
 public interface ServerConnectedController {
     boolean isWorkingOnline();
     boolean isWorkingOffline();
+    boolean isDisposed();
     boolean isValid();
 
     /**