+++ plugin/src/net/groboclown/idea/p4ic/v2/server/util/RemoteFileReader.java
@@ -105,14 +105,14 @@ public class RemoteFileReader {
                     encoding = es.get(0).getCharset();
                 }
                 if (encoding == null) {
-                    LOG.info("No known encoding for " + file + "; using utf-8");
+                    LOG.debug("No known encoding for " + file + "; using utf-8");
                     //alerts.addNotice(exec.getProject(),
                     //        P4Bundle.message("exception.load-file-encoding"),
                     //        new P4FileException(file),
                     //        file);
                     return new String(bytes, "utf-8");
                 }
-                LOG.info("reading " + file + " with encoding " + encoding);
+                LOG.debug("reading " + file + " with encoding " + encoding);
                 return new String(bytes, encoding);
             } catch (VcsException e) {
                 alerts.addWarning(exec.getProject(),